Move the modules
* MZ-700/emulator.js to MZ-700/MZ-700.js
* Z80/emulator.js to Z80/Z80.js
* MZ-700/sound.js to MZ-700/mz-beep.js
* MZ-700/mmio.js to MZ-700/mz-mmio.js
* Add MZ-700/mz-mmio-peripheral.js
* MZ-700/memory.js to MZ-700/mz700-memory.js
* MZ-700/monitor-rom,js to MZ-700/mz700-new-monitor.js
* MZ-700/worker.js to MZ-700/MZ-700-worker.js
* MZ-700/client.html to /emu.html and apply redirection
* MZ-700/index.js to MZ-700/MZ-700-emu-base.js
* MZ-700/index.css to MZ-700/MZ-700-emu.css
* MZ-700/client.js to MZ-700/MZ-700-emu.js
