+++ lib/PCG-700.js
@@ -46,6 +46,12 @@
     PCG700.SSW = 0x08;
     PCG700.ADDR = 0x07;
 
+    // Extends MmioPeripheral
+    const MmioPeripheral = require("./mz-mmio-peripheral.js");
+    Object.keys(MmioPeripheral.prototype).forEach( m => {
+        PCG700.prototype[m] = MmioPeripheral.prototype[m];
+    });
+
     PCG700.prototype.readMMIO = function(/* addr, value */) {}
 
     PCG700.prototype.writeMMIO = function(addr, value) {
