+++ lib/mzt-read-file.js
@@ -2,7 +2,7 @@
 (function() {
     "use strict";
     var fs = require('fs');
-    var MZ700 = require("../MZ-700/emulator.js");
+    var MZ700 = require("../MZ-700/MZ-700.js");
     module.exports = function (filename) {
         return new Promise(function(resolv, reject) {
             if(!filename) {
