+++ test/asm.js
@@ -1,7 +1,7 @@
 require("../lib/context.js");
 require('../lib/ex_number.js');
 var UnitTest = require("./UnitTest");
-var Z80 = require('../Z80/emulator.js');
+var Z80 = require('../Z80/Z80.js');
 var Z80_assemble = require('../Z80/assembler');
 var line_asm_test_pattern = [
     { code:[0x00], mnemonic:"NOP" },
