+++ test/rotate.js
@@ -1,6 +1,6 @@
 require("../lib/context.js");
 var Z80Tester = require('./Z80Tester.js');
-var Z80 = require('../Z80/emulator.js');
+var Z80 = require('../Z80/Z80.js');
 
 var tester = new Z80Tester();
 var cpu = new Z80();
