+++ test/search.js
@@ -1,7 +1,7 @@
 require("../lib/context.js");
 var UnitTest = require("./UnitTest");
 var Z80Tester = require('./Z80Tester.js');
-var Z80 = require('../Z80/emulator.js');
+var Z80 = require('../Z80/Z80.js');
 var cpu = new Z80();
 
 var tester = new Z80Tester();
